learning about the user, and detecting text
there are still major bugs in this part of the code which i will be fixing soon. i will also add an option to give god new words. i will also make it more interactive. thank you.
i also aded array type things, these are used to identify the text which will be used to throw a response. so far you can ask different questions. 
if there are any errors you see, please help and fix it! thanks.
if the "notgoodwords" array offends you please do not look at that part of the code. i had to use such words to make sure people are exploiting or making fun of the project.